added log and fixed minor issues

+++ mp3-convert.sh
@@ -28,22 +28,35 @@ done
 if [[ ! $bitrate ]]
 then
 	usage
+	exit 1
 fi
+
+PART=`date +"%s"`
+FILE="$HOME/mp3-convert_$PART.log"
 
 for f in *.mp3
 do
-	SONG=`id3ed -i "$f" | grep songname | sed 's/songname: //'`
-	ARTIST=`id3ed -i "$f" | grep artist | sed 's/artist: //'`
-	ALBUM=`id3ed -i "$f" | grep album | sed 's/album: //'`
-	YEAR=`id3ed -i "$f" | grep year | sed 's/year: //'`
-	TRACK=`id3ed -i "$f" | grep tracknum | sed 's/tracknum: //'`
+	if [ -f "$f" ]; then
+		echo "Starting new conversion" >>$FILE
+		echo "    Converting $f" >>$FILE
+		SONG=`id3ed -i "$f" | grep songname | sed 's/songname: //'`
+		ARTIST=`id3ed -i "$f" | grep artist | sed 's/artist: //'`
+		ALBUM=`id3ed -i "$f" | grep album | sed 's/album: //'`
+		YEAR=`id3ed -i "$f" | grep year | sed 's/year: //'`
+		TRACK=`id3ed -i "$f" | grep tracknum | sed 's/tracknum: //'`
 	
-	lame --mp3input -b $bitrate "$f" /tmp/"$f"
+		lame --mp3input -b $bitrate "$f" /tmp/"$f"
+		echo "    Finished converting $f" >>$FILE
+		mp3info -t "$SONG" -a "$ARTIST" -l "$ALBUM" -y "$YEAR" -n "$TRACK" /tmp/"$f"
 
-	mp3info -t "$SONG" -a "$ARTIST" -l "$ALBUM" -y "$YEAR" -n "$TRACK" /tmp/"$f"
-
-	rm "$f"	
+		rm "$f"
+		mv /tmp/"$f" "$FOLDER"
+		echo "    Restored from temp directory" >>$FILE
+		echo "Finished conversion" >>$FILE
+	else
+		echo "No mp3 files in current directory"
+	fi
 done
 
-mv /tmp/*.mp3 "$FOLDER"
+#mv /tmp/*.mp3 "$FOLDER"
